refactor: Move activities back into main package and make a utils package
