+++ code/app/src/main/java/com/hamidat/nullpointersapp/LoginActivity.java
@@ -1,4 +1,4 @@
-package com.hamidat.nullpointersapp.auth;
+package com.hamidat.nullpointersapp;
 
 import androidx.appcompat.app.AppCompatActivity;
