+++ code/app/src/main/java/com/hamidat/nullpointersapp/MainActivity.java
@@ -11,10 +11,6 @@ import androidx.core.graphics.Insets;
 import androidx.core.view.ViewCompat;
 import androidx.core.view.WindowInsetsCompat;
 
-import com.hamidat.nullpointersapp.auth.LoginActivity;
-import com.hamidat.nullpointersapp.auth.SignUpActivity;
-import com.hamidat.nullpointersapp.userProfile.ProfileActivity;
-
 public class MainActivity extends AppCompatActivity {
 
     @Override
